mop up more AI slop
