+++ rust/capture/tests/common/utils.rs
@@ -26,7 +26,7 @@ use tokio::time::timeout;
 use tracing::{info, warn, Level};
 
 use capture::config::{CaptureMode, Config, KafkaConfig};
-use capture::server::{self, serve};
+use capture::server::serve;
 use capture::setup;
 use common_continuous_profiling::ContinuousProfilingConfig;
 use limiters::redis::{QuotaResource, OVERFLOW_LIMITER_CACHE_KEY, QUOTA_LIMITER_CACHE_KEY};
@@ -168,8 +168,8 @@ impl ServerHandle {
             .with_shutdown_token(shutdown_token.clone())
             .build();
 
-        let handles = server::register_components(&mut manager, &config);
-        std::mem::forget(manager.monitor_background());
+        let handles = setup::register_components(&mut manager, &config);
+        let _monitor = manager.monitor_background();
         let components = setup::build_components(config, handles).await;
 
         tokio::spawn(async move { serve(listener, components).await });
